requirements.txt
arrow==1.2.3
asgiref==3.6.0
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
Flask==2.2.2
# GDAL==3.6.2
geos==0.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.1
Pillow==9.4.0
pip==23.0
proj==0.2.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
setuptools==60.2.0
six==1.16.0
spatialite==0.0.3
sqlparse==0.4.3
Werkzeug==2.2.2
wheel==0.37.1
